unsigned short type for ICON_SIZE and MAX_ICON_SIZE
(bzr r3684.3.6)

+++ plugins/compiztoolbox/include/compiztoolbox/compiztoolbox.h
@@ -161,7 +161,7 @@ class BaseSwitchWindow
 	CompWindow       *window;
 };
 
-extern const unsigned int ICON_SIZE;
-extern const unsigned int MAX_ICON_SIZE;
+extern const unsigned short ICON_SIZE;
+extern const unsigned short MAX_ICON_SIZE;
 
 #endif
